+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-CV1oMcA6.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW--m8lqAn5.js";import{_ as a}from"./mermaid.core-BRlXsNXe.js";import"./chunk-5VM5RSS4-DPnc7_al.js";import"./chunk-XXDRQBXY-DdE8UFz5.js";import"./chunk-POPQ4Y6H-DKrvEy2f.js";import"./chunk-F27PBJKO-CEHmKYhA.js";import"./channel-BnoujDFM.js";import"./index-Dz_N_J9m.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};